requirements.txt
#-e .
aws-cdk.core

aws-cdk.aws_lambda
aws-cdk.aws_dynamodb

boto3
botocore